+++ Consul.Test/Properties/AssemblyInfo.cs
@@ -36,4 +36,4 @@ using Xunit;
 [assembly: AssemblyVersion("1.0.0.0")]
 [assembly: AssemblyFileVersion("1.0.0.0")]
 //[assembly: CollectionBehavior(DisableTestParallelization = false)]
-[assembly: CollectionBehavior(MaxParallelThreads = 800)]
+[assembly: CollectionBehavior(MaxParallelThreads = 1024)]